add dtype to bbox unit test
