+++ tests/test_data/test_pipelines/test_transform/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) OpenMMLab. All rights reserved.
+from .utils import construct_toy_data, create_random_bboxes
+
+__all__ = ['create_random_bboxes', 'construct_toy_data']
