+++ tests/test_data/test_pipelines/test_transform/utils.py
@@ -1,0 +1,68 @@
+# Copyright (c) OpenMMLab. All rights reserved.
+import numpy as np
+
+from mmdet.core.mask import BitmapMasks, PolygonMasks
+
+
+def _check_fields(results, pipeline_results, keys):
+    for key in keys:
+        if isinstance(results[key], (BitmapMasks, PolygonMasks)):
+            assert np.equal(results[key].to_ndarray(),
+                            pipeline_results[key].to_ndarray()).all()
+        else:
+            assert np.equal(results[key], pipeline_results[key]).all()
+
+
+def check_result_same(results, pipeline_results):
+    # check image
+    _check_fields(results, pipeline_results,
+                  results.get('img_fields', ['img']))
+    # check bboxes
+    _check_fields(results, pipeline_results, results.get('bbox_fields', []))
+    # check masks
+    _check_fields(results, pipeline_results, results.get('mask_fields', []))
+    # check segmentations
+    _check_fields(results, pipeline_results, results.get('seg_fields', []))
+    # check gt_labels
+    if 'gt_labels' in results:
+        assert np.equal(results['gt_labels'],
+                        pipeline_results['gt_labels']).all()
+
+
+def construct_toy_data(poly2mask=True):
+    img = np.array([[1, 2, 3, 4], [5, 6, 7, 8]], dtype=np.uint8)
+    img = np.stack([img, img, img], axis=-1)
+    results = dict()
+    # image
+    results['img'] = img
+    results['img_shape'] = img.shape
+    results['img_fields'] = ['img']
+    # bboxes
+    results['bbox_fields'] = ['gt_bboxes', 'gt_bboxes_ignore']
+    results['gt_bboxes'] = np.array([[0., 0., 2., 1.]], dtype=np.float32)
+    results['gt_bboxes_ignore'] = np.array([[2., 0., 3., 1.]],
+                                           dtype=np.float32)
+    # labels
+    results['gt_labels'] = np.array([1], dtype=np.int64)
+    # masks
+    results['mask_fields'] = ['gt_masks']
+    if poly2mask:
+        gt_masks = np.array([[0, 1, 1, 0], [0, 1, 0, 0]],
+                            dtype=np.uint8)[None, :, :]
+        results['gt_masks'] = BitmapMasks(gt_masks, 2, 4)
+    else:
+        raw_masks = [[np.array([0, 0, 2, 0, 2, 1, 0, 1], dtype=np.float)]]
+        results['gt_masks'] = PolygonMasks(raw_masks, 2, 4)
+    # segmentations
+    results['seg_fields'] = ['gt_semantic_seg']
+    results['gt_semantic_seg'] = img[..., 0]
+    return results
+
+
+def create_random_bboxes(num_bboxes, img_w, img_h):
+    bboxes_left_top = np.random.uniform(0, 0.5, size=(num_bboxes, 2))
+    bboxes_right_bottom = np.random.uniform(0.5, 1, size=(num_bboxes, 2))
+    bboxes = np.concatenate((bboxes_left_top, bboxes_right_bottom), 1)
+    bboxes = (bboxes * np.array([img_w, img_h, img_w, img_h])).astype(
+        np.float32)
+    return bboxes
